+++ src/main/java/com/gg/server/domain/tournament/data/Tournament.java
@@ -20,11 +20,11 @@ import lombok.*;
 @ToString
 public class Tournament extends BaseTimeEntity {
     // 토너먼트 참가자 수 => 현재는 8강 고정
-    public static final long ALLOWED_JOINED_NUMBER = 8;
+    public static final int ALLOWED_JOINED_NUMBER = 8;
     // 토너먼트 최소 시작 날짜 (n일 후)
-    public static final long ALLOWED_MINIMAL_START_DAYS = 2;
+    public static final int ALLOWED_MINIMAL_START_DAYS = 2;
     // 토너먼트 최소 진행 시간 (n시간)
-    public static final long MINIMUM_TOURNAMENT_DURATION = 2;
+    public static final int MINIMUM_TOURNAMENT_DURATION = 2;
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
